tested autonomous paths and stack intake
