+++ 2024/opmodes/sensortests/DistanceSensorTest.java
@@ -1,4 +1,4 @@
-package opmodes.stack;
+package opmodes.sensortests;
 
 import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.FLIP_DOWN;
 import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.FLIP_UP;
