+++ 2024/opmodes/stack/RRAutoCSBlueLeftV2Stack.java
@@ -1,5 +1,6 @@
 package opmodes.stack;
 
+import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.BLOCK_PIXELS;
 import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.FLIP_DOWN;
 import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.FLIP_UP;
 import static org.firstinspires.ftc.teamcode.drive.CenterstageSampleMecanumDrive.RELEASE_PIXELS;
@@ -31,7 +32,6 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
     public static double DISTANCE = 30; // in
 
     private final double BLOCK_NOTHING = 0.05;
-    private final double BLOCK_BOTH = 0.8;
 
     private final double PROP_DIST = 10; // cm
 
@@ -54,6 +54,7 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
         TrajectorySequence toSpikes = drive.trajectorySequenceBuilder(startPose)
                 // APPROACHING SPIKES
                 .lineToLinearHeading(new Pose2d(12, 29, Math.toRadians(90)))
+                .waitSeconds(0.5)
                 .build();
         TrajectorySequence leftSpike = drive.trajectorySequenceBuilder(toSpikes.end())
                 // LEFT SPIKE PATH
@@ -92,6 +93,7 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
                     while (drive.colorSensor.red()<300 && drive.colorSensor.green()<400 && drive.colorSensor.blue()<300) {
                     // color is black --> intake
                         drive.intake.setPower(-0.9);
+                        sleep(1000); // FIXME test this to see if the delay works. goal is to only stop intaking if we get 2 pixels, but just when firts pixel bounces up
                     }
                     // after color is not black (meaning it's yellow), stop intaking
                     drive.intake.setPower(0);})
@@ -101,8 +103,8 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
                 .UNSTABLE_addTemporalMarkerOffset(4, () -> {drive.intake.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.intake.setPower(-0.9);})
                 .setTangent(0)
-                .splineToConstantHeading(new Vector2d(-34,55),Math.toRadians(0))
-                .lineToLinearHeading(new Pose2d(30,55,Math.toRadians(180)))
+                .splineToConstantHeading(new Vector2d(-34,55.5),Math.toRadians(0))
+                .lineToLinearHeading(new Pose2d(30,55.5,Math.toRadians(180)))
                 .UNSTABLE_addTemporalMarkerOffset(0.8, () -> {drive.box.setPosition(FLIP_UP);})
                 .UNSTABLE_addTemporalMarkerOffset(0.7, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(0.4);})
@@ -133,14 +135,14 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
                 .waitSeconds(0.5)
                 .UNSTABLE_addTemporalMarkerOffset(0.6, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(-0.4);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.pixelRelease.setPosition(BLOCK_BOTH);})
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.pixelRelease.setPosition(BLOCK_PIXELS);})
                 .setTangent(90)
                 .splineToConstantHeading(new Vector2d(30,55),Math.toRadians(180))
                 .waitSeconds(0.5)
                 .lineToLinearHeading(new Pose2d(-34,55,Math.toRadians(180)))
-                .splineToConstantHeading(new Vector2d(-51,32),Math.toRadians(180))
+                .splineToConstantHeading(new Vector2d(-51,33),Math.toRadians(180))
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linkage.setPosition(0.5);})
                 .forward(5)
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linkage.setPosition(0.2);})
                 .waitSeconds(1)
                 // intake stack pixels
                 .UNSTABLE_addTemporalMarkerOffset(3, () -> {drive.intake.setPower(0);})
@@ -154,15 +156,14 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
                 .UNSTABLE_addTemporalMarkerOffset(7, () -> {drive.intake.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.intake.setPower(-0.9);})
                 .setTangent(0)
-                .splineToConstantHeading(new Vector2d(-34,53),Math.toRadians(0))
-                .lineToLinearHeading(new Pose2d(30,53,Math.toRadians(180)))
+                .splineToConstantHeading(new Vector2d(-34,54),Math.toRadians(0))
+                .lineToLinearHeading(new Pose2d(30,54,Math.toRadians(180)))
                 .setTangent(0)
                 .UNSTABLE_addTemporalMarkerOffset(0.8, () -> {drive.box.setPosition(FLIP_UP);})
                 .UNSTABLE_addTemporalMarkerOffset(0.7, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(0.4);})
-                .splineToConstantHeading(new Vector2d(53, 34), Math.toRadians(0))
-                .waitSeconds(1)
                 .UNSTABLE_addTemporalMarkerOffset(6, () -> {drive.pixelRelease.setPosition(RELEASE_PIXELS);})
+                .splineToConstantHeading(new Vector2d(53, 34), Math.toRadians(0))
                 .build();
         TrajectorySequence rightSpike = drive.trajectorySequenceBuilder(toSpikes.end())
                 // RIGHT SPIKE PATH
@@ -177,47 +178,52 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
                 .UNSTABLE_addTemporalMarkerOffset(0.7, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(0.4);})
                 .waitSeconds(0.5)
-                .lineToLinearHeading(new Pose2d(55, 31, Math.toRadians(180)))
-                .UNSTABLE_addTemporalMarkerOffset(0.6, () -> {drive.linearLift.setPower(0);})
+                .lineToLinearHeading(new Pose2d(55, 30.4, Math.toRadians(180)))
+                .UNSTABLE_addTemporalMarkerOffset(0.4, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(-0.4);})
                 // * deploy yellow pixel
                 .waitSeconds(1)
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.pixelRelease.setPosition(RELEASE_PIXELS);})
                 .waitSeconds(1)
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.box.setPosition(FLIP_DOWN);})
-                .waitSeconds(0.5)
-                .UNSTABLE_addTemporalMarkerOffset(0.6, () -> {drive.linearLift.setPower(0);})
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.pixelRelease.setPosition(BLOCK_PIXELS);})
+                .UNSTABLE_addTemporalMarkerOffset(0.3, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(-0.4);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.pixelRelease.setPosition(BLOCK_BOTH);})
                 .setTangent(90)
                 .splineToConstantHeading(new Vector2d(30,55),Math.toRadians(180))
                 .waitSeconds(0.5)
                 .lineToLinearHeading(new Pose2d(-34,55,Math.toRadians(180)))
-                .splineToConstantHeading(new Vector2d(-51,33),Math.toRadians(180))
-                .forward(5)
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linkage.setPosition(0.2);})
-                .waitSeconds(1)
+                .splineToConstantHeading(new Vector2d(-48,48),Math.toRadians(180))
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linkage.setPosition(0.5);})
+                .lineToLinearHeading(new Pose2d(-57.8, 38,Math.toRadians(207)))
                 // intake stack pixels
-                .UNSTABLE_addTemporalMarkerOffset(3, () -> {drive.intake.setPower(0);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.intake.setPower(-0.9);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linkage.setPosition(0.25);})
-                .waitSeconds(1.5)
-                .back(3)
-                .UNSTABLE_addTemporalMarkerOffset(4, () -> {drive.intake.setPower(0);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.intake.setPower(0.9);})
-                .waitSeconds(2)
-                .UNSTABLE_addTemporalMarkerOffset(7, () -> {drive.intake.setPower(0);})
-                .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.intake.setPower(-0.9);})
-                .setTangent(0)
-                .splineToConstantHeading(new Vector2d(-34,55.5),Math.toRadians(0))
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {
+                    while (drive.colorSensor.red()<300 && drive.colorSensor.green()<400 && drive.colorSensor.blue()<300) {
+                        // color is black --> intake
+                        drive.intake.setPower(-0.9);
+                    }
+                    // after color is not black (meaning it's yellow), stop intaking
+                    drive.intake.setPower(0);
+                })
                 .waitSeconds(0.5)
-                .lineToLinearHeading(new Pose2d(30,55.5,Math.toRadians(180)))
+                // outtake while driving backwards
+                .UNSTABLE_addTemporalMarkerOffset(3, () -> {
+                    drive.intake.setPower(0);
+                })
+                .UNSTABLE_addTemporalMarkerOffset(0, () -> {
+                    drive.intake.setPower(0.9);
+                })
+                .setTangent(0)
+                .splineToConstantHeading(new Vector2d(-34,54),Math.toRadians(0))
+                .turn(Math.toRadians(-36))
+                .waitSeconds(0.5)
+                .lineToLinearHeading(new Pose2d(30,54,Math.toRadians(180)))
                 .setTangent(0)
                 .UNSTABLE_addTemporalMarkerOffset(0.8, () -> {drive.box.setPosition(FLIP_UP);})
-                .UNSTABLE_addTemporalMarkerOffset(0.7, () -> {drive.linearLift.setPower(0);})
+                .UNSTABLE_addTemporalMarkerOffset(0.45, () -> {drive.linearLift.setPower(0);})
                 .UNSTABLE_addTemporalMarkerOffset(0, () -> {drive.linearLift.setPower(0.4);})
-                .UNSTABLE_addTemporalMarkerOffset(6, () -> {drive.pixelRelease.setPosition(RELEASE_PIXELS);})
-                .splineToConstantHeading(new Vector2d(52, 34), Math.toRadians(0))
+                .UNSTABLE_addTemporalMarkerOffset(8, () -> {drive.pixelRelease.setPosition(RELEASE_PIXELS);})
+                .splineToConstantHeading(new Vector2d(51.5, 34), Math.toRadians(0))
                 .build();
         waitForStart();
 
@@ -225,15 +231,20 @@ public class RRAutoCSBlueLeftV2Stack extends LinearOpMode {
 
         drive.followTrajectorySequence(toSpikes);
 
-
         if (detectProp() == "left") {
             drive.followTrajectorySequence(leftSpike);
+            telemetry.addData("dist sensor 1 range", String.format("%.01f cm", drive.distanceSensor1.getDistance(DistanceUnit.CM)));
+            telemetry.addData("dist sensor 2 range", String.format("%.01f cm", drive.distanceSensor2.getDistance(DistanceUnit.CM)));
         }
         else if (detectProp() == "center") {
             drive.followTrajectorySequence(centerSpike);
+            telemetry.addData("dist sensor 1 range", String.format("%.01f cm", drive.distanceSensor1.getDistance(DistanceUnit.CM)));
+            telemetry.addData("dist sensor 2 range", String.format("%.01f cm", drive.distanceSensor2.getDistance(DistanceUnit.CM)));
         }
         else {
             drive.followTrajectorySequence(rightSpike);
+            telemetry.addData("dist sensor 1 range", String.format("%.01f cm", drive.distanceSensor1.getDistance(DistanceUnit.CM)));
+            telemetry.addData("dist sensor 2 range", String.format("%.01f cm", drive.distanceSensor2.getDistance(DistanceUnit.CM)));
         }
